Driver updates

+++ driver.py
@@ -3,6 +3,7 @@ import subprocess
 import os
 import numpy as np
 from time import sleep,time,localtime,strftime,monotonic
+from datetime import datetime
 # from multiprocessing import Process, Queue, Array
 import multiprocessing as mp
 from multiprocessing.shared_memory import SharedMemory
@@ -43,17 +44,21 @@ def throttle(target_fps, start_time):
     else:
         return False
     
-def read_frames(cap: cv2.VideoCapture, shmem_name: str, cur_frame_idx: mp.Value, target_fps: int):
+def read_frames(cap: cv2.VideoCapture, shmem_name: str, cur_frame_idx: mp.Value, target_fps: int, ret_queue: mp.Queue, frame_drop: mp.Queue):
     existing_shm = SharedMemory(name=shmem_name)
     shared_array = np.ndarray((height, width, 3), dtype=np.uint8, buffer=existing_shm.buf)
+    total_frames_set = set() # Set to keep track of frame dropping
 
     _ = cap.set(cv2.CAP_PROP_POS_FRAMES, 0)
     total_frames = 0
+    read_start_time = monotonic()
     while True:
         now = monotonic()
         ret, frame = cap.read()
         if not ret:
             print('============ Finished video ============')
+            ret_queue.put(total_frames / (monotonic() - read_start_time))
+            frame_drop.put(total_frames_set)
             cur_frame_idx.value = -1 # Signal that video has ended
             return
         with cur_frame_idx.get_lock():
@@ -61,13 +66,17 @@ def read_frames(cap: cv2.VideoCapture, shmem_name: str, cur_frame_idx: mp.Value,
             cur_frame_idx.value = total_frames
 
         throttle(target_fps, now)
+        total_frames_set.add(total_frames)
         total_frames += 1
 
-def filter_frames(diff_processor, filter_shmem_name: str, filter_frame_idx: mp.Value, encoding_shmem_name: str, encoding_frame_idx: mp.Value, width: int, height: int):
+def filter_frames(diff_processor, filter_shmem_name: str, filter_frame_idx: mp.Value, encoding_shmem_name: str, encoding_frame_idx: mp.Value, width: int, height: int, frame_drop: mp.Queue):
     filter_shm = SharedMemory(name=filter_shmem_name)
     encoding_shm = SharedMemory(name=encoding_shmem_name)
     filter_shared_array = np.ndarray((height, width, 3), dtype=np.uint8, buffer=filter_shm.buf)
     encoding_shared_array = np.ndarray((height, width, 3), dtype=np.uint8, buffer=encoding_shm.buf)
+
+    filtered_input_set = set() # Set to keep track of frame dropping
+    filter_output_set = set()
 
     max_dis = 0
     min_dis = 99999999
@@ -88,6 +97,8 @@ def filter_frames(diff_processor, filter_shmem_name: str, filter_frame_idx: mp.V
             frame = filter_shared_array.copy()
             frame_idx = filter_frame_idx.value
         if frame_idx == -1:
+            frame_drop.put(filtered_input_set)
+            frame_drop.put(filter_output_set)
             filter_frame_idx.value = -2
             encoding_frame_idx.value = -1
             # print(f'MAX DIFFERENCE {max_dis}')
@@ -99,6 +110,7 @@ def filter_frames(diff_processor, filter_shmem_name: str, filter_frame_idx: mp.V
 
         feat = diff_processor.get_frame_feature(frame)
         dis = diff_processor.cal_frame_diff(feat, prev_feat)
+        filtered_input_set.add(frame_idx)
 
         # print(prev_idx, frame_idx)
         # print(f'difference: {dis}')
@@ -107,18 +119,21 @@ def filter_frames(diff_processor, filter_shmem_name: str, filter_frame_idx: mp.V
 
         if dis > diff_processor.thresh:
             prev_feat = feat
-            prev_idx = frame_idx
+            prev_idx = frame_idx # TODO: Move this out of if statement?
             with encoding_frame_idx.get_lock():
                 encoding_frame_idx.value = frame_idx
                 encoding_shared_array[:,:,:] = frame
+            filter_output_set.add(frame_idx)
 
-def encode_frames(encoding_shmem_name: str, encoding_frame_idx: mp.Value, bitrate: int, fps: int, width: int, height: int, save_dir: str):
+def encode_frames(encoding_shmem_name: str, encoding_frame_idx: mp.Value, bitrate: int, fps: int, width: int, height: int, save_dir: str, frame_drop: mp.Queue):
     encoder = ffenc(width, height, fps)
     decoder = ffdec()
     encoder.change_settings(bitrate, fps)
 
     encoding_shm = SharedMemory(name=encoding_shmem_name)
     encoding_shared_array = np.ndarray((height, width, 3), dtype=np.uint8, buffer=encoding_shm.buf)
+
+    encoded_frames_set = set() # Set to keep track of frame dropping
 
     prev_frame_idx = -2 # Use -2 to avoid conflict with -1 signaling termination
 
@@ -128,16 +143,19 @@ def encode_frames(encoding_shmem_name: str, encoding_frame_idx: mp.Value, bitrat
             frame = encoding_shared_array.copy()
         # print(f'encoding_frame_idx.value: {enc_frame_idx}')
         if enc_frame_idx == -1:
+            frame_drop.put(encoded_frames_set)
             encoding_frame_idx.value = -2
             return
         elif prev_frame_idx == enc_frame_idx:
-            sleep(0.05)
+            sleep(0.01)
             continue
         prev_frame_idx = enc_frame_idx
         # print(f'Encode: Saving frame {enc_frame_idx}')
         encoded_frame = encoder.process_frame(frame)
         decoded_frame = decoder.process_frame(encoded_frame)
         decoded_frame = cv2.cvtColor(decoded_frame, cv2.COLOR_RGB2BGR)
+
+        encoded_frames_set.add(enc_frame_idx)
 
         # Save raw image
         filename = os.path.join(save_dir, f'frame{enc_frame_idx}.npy')
@@ -177,10 +195,12 @@ def class2str(cls):
     else:
         raise Exception('Unknown class')
     
-def read_energy(TC66, out_file, start, energy_readings, interval=1):
-    with open(out_file,'w') as f:
+def read_energy(TC66, start, energy_readings, out_file='TC66_'+strftime('%Y%m%d%H%M%S',localtime())+'.csv', interval=1):
+    if out_file:
+        f = open(out_file,'w')
         f.write('Time[S],Volt[V],Current[A],Power[W]\n')
 
+    try:
         while True:
             now = monotonic()-start
             pd = TC66.Poll()
@@ -189,7 +209,8 @@ def read_energy(TC66, out_file, start, energy_readings, interval=1):
                 pd.Volt, 
                 pd.Current,
                 pd.Power)
-            f.write(s+'\n')
+            if out_file:
+                f.write(s+'\n')
 
             energy_readings.put(pd.Power)
 
@@ -197,6 +218,9 @@ def read_energy(TC66, out_file, start, energy_readings, interval=1):
             elapsed = (monotonic()-start) - now
             if elapsed < interval:
                 sleep(interval - elapsed)
+    except KeyboardInterrupt:
+        if out_file:
+            f.close()
 
 def get_average_energy(energy_readings):
     power_sum = 0
@@ -211,7 +235,8 @@ def get_average_energy(energy_readings):
 if __name__ == '__main__':
     print('Running')
 
-    LOG_FILE = './1_5ghz.csv'
+    current_time = datetime.now()
+    LOG_FILE = f'./test-{current_time.strftime("%Y%m%d%H%M%S")}.csv'
     VIDEO = './videos/ny_driving.nut'
     FREQUENCIES = [1500000, 1800000, 2100000, 2400000]
     FILTERS = [PixelDiff, AreaDiff, EdgeDiff]
@@ -220,9 +245,9 @@ if __name__ == '__main__':
     # Batch 3: [0.7, 0.8, 0.9]
     # THRESHOLDS = [0.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9]
 
-    PIXEL_THRESHOLDS = [0.0, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1 ]
-    AREA_THRESHOLDS = [0.0, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1 ]
-    EDGE_THRESHOLDS = [0.0, 0.0004, 0.0008, 0.0012, 0.0016, 0.002, 0.0024, 0.0028, 0.0032, 0.0036, 0.004]
+    PIXEL_THRESHOLDS = [0.00, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.10 ]
+    AREA_THRESHOLDS = [0.00, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.10 ]
+    EDGE_THRESHOLDS = [0.00, 0.0004, 0.0008, 0.0012, 0.0016, 0.0020, 0.0024, 0.0028, 0.0032, 0.0036, 0.0040]
 
     FRAME_BITRATES = [100, 400, 700, 1000, 1300, 1600, 1900, 2100, 2400, 2700, 3000] # kbps
 
@@ -242,31 +267,36 @@ if __name__ == '__main__':
 
     total_exp_start_time = monotonic()
 
-    filter_frame_idx = mp.Value('i')
-    encoding_frame_idx = mp.Value('i')
-    filter_frame_idx.value = -2
-    encoding_frame_idx.value = -2
-
-    filter_shm = SharedMemory(create=True, size=height*width*3)
-    encoding_shm = SharedMemory(create=True, size=height*width*3)
     # encoding_queue = mp.Queue()
     energy_readings = mp.Queue()
 
     mp.Process(target=read_energy, args=(TC66,
-                                      'TC66_'+strftime('%Y%m%d%H%M%S',localtime())+'.csv',
                                       total_exp_start_time, energy_readings),
-                                      kwargs = {'interval': 0.25}).start()
+                                      kwargs = {'out_file': None,
+                                                'interval': 0.25}).start()
 
     # gt_dir = os.path.join(os.path.dirname(__file__), 'ground-truth')
     # generate_ground_truth(cap, width, height, 3000, 30, gt_dir)
 
     try:
-        for frequency in [1500000]:
+        for frequency in [1800000]:
 
             set_cpu_freq(frequency)
 
+            filter_shm = SharedMemory(create=True, size=height*width*3)
+            encoding_shm = SharedMemory(create=True, size=height*width*3)
+
+            filter_frame_idx = mp.Value('i')
+            encoding_frame_idx = mp.Value('i')
+            filter_frame_idx.value = -2
+            encoding_frame_idx.value = -2
+
+            return_values = mp.Queue()
+            frame_drop = mp.Queue() # For frame index sets at each step
+
             for filter in FILTERS:
-                
+            # for filter in [AreaDiff]:
+
                 if filter == PixelDiff:
                     thresholds = PIXEL_THRESHOLDS
                 elif filter == AreaDiff:
@@ -277,21 +307,23 @@ if __name__ == '__main__':
                     raise ValueError('ERROR: Filter not recognized')
 
                 for threshold in thresholds:
+                # for threshold in [0.00]:
 
                     diff_processor = filter(thresh=threshold)
 
                     for frame_bitrate in FRAME_BITRATES:
+                    # for frame_bitrate in [3000] * 2:
                         bitrate = frame_bitrate * fps
-                        save_dir = os.path.join(os.path.dirname(__file__), 'flashdrive', '1.5', f'{frequency / 1_000_000}-{class2str(filter)}-{threshold}-{frame_bitrate}')
+                        save_dir = os.path.join(os.path.dirname(__file__), 'flashdrive', f'{frequency / 1_000_000}-{class2str(filter)}-{threshold:.4f}-{frame_bitrate}')
                         # save_dir = os.path.join('/home', 'bencivjan', 'Desktop', 'flashdrive', 'batch1', f'{frequency / 1_000_000}-{class2str(filter)}-{threshold}-{frame_bitrate}')
                         os.makedirs(save_dir, exist_ok=True)
 
                         cur_exp_start_time = monotonic()
 
                         # num_frames = encode_video(cap, diff_processor, encoder, decoder, save_dir, target_fps, total_exp_start_time)
-                        read_frames_pid = mp.Process(target=read_frames, args=(cap, filter_shm.name, filter_frame_idx, target_fps))
-                        filter_frames_pid = mp.Process(target=filter_frames, args=(diff_processor, filter_shm.name, filter_frame_idx, encoding_shm.name, encoding_frame_idx,  width, height))
-                        encode_frames_pid = mp.Process(target=encode_frames, args=(encoding_shm.name, encoding_frame_idx, bitrate, fps, width, height, save_dir))
+                        read_frames_pid = mp.Process(target=read_frames, args=(cap, filter_shm.name, filter_frame_idx, target_fps, return_values, frame_drop))
+                        filter_frames_pid = mp.Process(target=filter_frames, args=(diff_processor, filter_shm.name, filter_frame_idx, encoding_shm.name, encoding_frame_idx,  width, height, frame_drop))
+                        encode_frames_pid = mp.Process(target=encode_frames, args=(encoding_shm.name, encoding_frame_idx, bitrate, fps, width, height, save_dir, frame_drop))
 
                         get_average_energy(energy_readings) # Clear out readings collected before experiment
 
@@ -303,6 +335,21 @@ if __name__ == '__main__':
                         filter_frames_pid.join()
                         encode_frames_pid.join()
 
+                        real_fps = return_values.get()
+
+                        total_frames_set = frame_drop.get()
+                        filter_input_set = frame_drop.get()
+                        filter_output_set = frame_drop.get()
+                        encoded_frames_set = frame_drop.get()
+                        filter_input_dropped_frames = total_frames_set - filter_input_set
+                        filtered_frames = filter_input_set - filter_output_set
+                        encoder_dropped_frames = filter_output_set - encoded_frames_set
+                        total_dropped_frames = total_frames_set - encoded_frames_set
+                        print(f'Filter input dropped frames: {len(filter_input_dropped_frames)}')
+                        print(f'Filtered frames: {len(filtered_frames)}')
+                        print(f'Encoder dropped frames: {len(encoder_dropped_frames)}')
+                        print(f'Total dropped frames: {len(total_dropped_frames)}')
+
                         cur_exp_end_time = monotonic()
                         tot_time = cur_exp_end_time - cur_exp_start_time
                         average_energy = get_average_energy(energy_readings)
@@ -313,16 +360,27 @@ if __name__ == '__main__':
                                 start_time = cur_exp_start_time - total_exp_start_time
                                 end_time = cur_exp_end_time - total_exp_start_time
 
-                                file.write(f'{freq_ghz},{filter_str},{threshold},{frame_bitrate},{fps:.3f},{start_time:.3f},{end_time:.3f},{average_energy}\n')
-
+                                file.write(f'{freq_ghz},{filter_str},{threshold:4f},{frame_bitrate},{real_fps:.3f},{start_time:.3f},{end_time:.3f},{average_energy}\n')
+                        
                         sleep(10)
+            
+            # Clean up shared memory
+            filter_shm.close()
+            filter_shm.unlink()
+            encoding_shm.close()
+            encoding_shm.unlink()
+            filter_shm = None
+            encoding_shm = None
+
     except Exception as e:
         print(f'ERROR: {e}')
         with open('errors.out', 'w') as f:
-            f.write(e)
+            f.write(str(e))
     finally:
         # Clean up shared memory
-        filter_shm.close()
-        filter_shm.unlink()
-        encoding_shm.close()
-        encoding_shm.unlink()
+        if filter_shm:
+            filter_shm.close()
+            filter_shm.unlink()
+        if encoding_shm:
+            encoding_shm.close()
+            encoding_shm.unlink()
